+++ messages/ru/yii-validator.php
@@ -136,4 +136,23 @@ return [
     'The allowed types are integer, float and string.' => 'Разрешённые типы: integer, float и string.',
     '{Attribute} must be no less than {min}.' => '{Attribute} должно быть не меньше {min}.',
     '{Attribute} must be no greater than {max}.' => '{Attribute} должно быть не больше {max}.',
+
+    /**
+     * @see \Yiisoft\Validator\Rule\Date\Date
+     * @see \Yiisoft\Validator\Rule\Date\DateTime
+     * @see \Yiisoft\Validator\Rule\Date\Time
+     */
+    'The value must be no early than {limit}.' => 'Значение должно быть не ранее {limit}.',
+    'The value must be no late than {limit}.' => 'Значение должно быть не позднее {limit}.',
+
+    /**
+     * @see \Yiisoft\Validator\Rule\Date\Date
+     * @see \Yiisoft\Validator\Rule\Date\DateTime
+     */
+    'Invalid date value.' => 'Некорректное значение даты.',
+
+    /**
+     * @see \Yiisoft\Validator\Rule\Date\Time
+     */
+    'Invalid time value.' => 'Некорректное значение времени.',
 ];