+++ src/Exception/UnexpectedRuleException.php
@@ -36,28 +36,22 @@ use Throwable;
  */
 final class UnexpectedRuleException extends InvalidArgumentException
 {
+    /**
+     * @param string|string[] $expectedClassName Expected class name(s) of a rule.
+     * @param object $actualObject An actual given object that's not an instance of `$expectedClassName`.
+     * @param int $code The Exception code.
+     * @param Throwable|null $previous The previous throwable used for the exception chaining.
+     */
     public function __construct(
-        /**
-         * @var string Expected class name of a rule.
-         */
-        string $expectedClassName,
-        /**
-         * @var object An actual given object that's not an instance of `$expectedClassName`.
-         */
+        string|array $expectedClassName,
         object $actualObject,
-        /**
-         * @var int The Exception code.
-         */
         int $code = 0,
-        /**
-         * @var Throwable|null The previous throwable used for the exception chaining.
-         */
         ?Throwable $previous = null,
     ) {
         parent::__construct(
             sprintf(
                 'Expected "%s", but "%s" given.',
-                $expectedClassName,
+                implode('", "', (array) $expectedClassName),
                 $actualObject::class
             ),
             $code,