+++ src/Rule/Image/ImageInfo.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Validator\Rule\Image;
+
+final class ImageInfo
+{
+    public function __construct(
+        private int $width,
+        private int $height,
+    ) {
+    }
+
+    public function getWidth(): int
+    {
+        return $this->width;
+    }
+
+    public function getHeight(): int
+    {
+        return $this->height;
+    }
+}