+++ tests/Rule/Base/RuleTestCase.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace Yiisoft\Validator\Tests\Rule\Base;
 
 use PHPUnit\Framework\TestCase;
+use Yiisoft\Validator\Result;
+use Yiisoft\Validator\RuleHandlerResolver\SimpleRuleHandlerContainer;
 use Yiisoft\Validator\RuleInterface;
 use Yiisoft\Validator\Validator;
 
@@ -15,9 +17,15 @@ abstract class RuleTestCase extends TestCase
     /**
      * @dataProvider dataValidationPassed
      */
-    public function testValidationPassed(mixed $data, ?array $rules = null): void
-    {
-        $result = (new Validator())->validate($data, $rules);
+    public function testValidationPassed(
+        mixed $data,
+        array|RuleInterface|null $rules = null,
+        ?array $ruleHandlers = null
+    ): void {
+        $validator = new Validator(
+            ruleHandlerResolver: $ruleHandlers === null ? null : new SimpleRuleHandlerContainer($ruleHandlers)
+        );
+        $result = $validator->validate($data, $rules);
 
         $this->assertSame([], $result->getErrorMessagesIndexedByPath());
     }
@@ -30,11 +38,26 @@ abstract class RuleTestCase extends TestCase
     public function testValidationFailed(
         mixed $data,
         array|RuleInterface|null $rules,
-        array $errorMessagesIndexedByPath
+        array $errorMessagesIndexedByPath,
+        ?array $ruleHandlers = null
     ): void {
-        $result = (new Validator())->validate($data, $rules);
+        $validator = new Validator(
+            ruleHandlerResolver: $ruleHandlers === null ? null : new SimpleRuleHandlerContainer($ruleHandlers)
+        );
+        $result = $validator->validate($data, $rules);
 
         $this->assertFalse($result->isValid());
-        $this->assertSame($errorMessagesIndexedByPath, $result->getErrorMessagesIndexedByPath());
+        $this->assertSame($errorMessagesIndexedByPath, $this->getErrorMessages($result));
+    }
+
+    private function getErrorMessages(Result $result): array
+    {
+        return array_map(
+            static fn(array $errors) => array_map(
+                static fn(string $error) => str_replace(' ', ' ', $error),
+                $errors
+            ),
+            $result->getErrorMessagesIndexedByPath(),
+        );
     }
 }